avoid configure substitutions in _SOURCES variables
Automake manual says:
"You can’t put a configure substitution (e.g., ‘@FOO@’ or ‘$(FOO)’
where FOO is defined via AC_SUBST) into a _SOURCES variable. The
reason for this is a bit hard to explain, but suffice to say that it
simply won’t work. Automake will give an error if you try to do this."

Well, okay.