fix check for the deh_bexptrs[i].lookup being NULL
The deh_bexptrs[i].lookup is a pointer to a char array and is never
NULL. Check for the actual code pointer being NULL instead. This fixes
a crash with Dehacked patches pointing to non-existent code pointers,
e.g. the one embedded into SihR2fix.wad.

Thanks Catpho for the report.