composite all textures, whether single-patched or not  (#97)
* composite all textures, whether single-patched or not

With this commit, Crispy now composes all textures, whether they are
single-patched or not. It generates to composites per texture, one for
rendering mid-textures on 2S wall which may contain translucent areas,
and another one for rendering textures on 1S walls and the sky,
respectively.

No column is ever read directly from a patch lump anymore. This keeps
the Medusa bug fixed but also fixes graphical glitches with e.g.
single-patched textures that are shorter than the patch, etc.

* initialize opaque texture column offset earlier

* compose tall patch compliant texture columns

Once we pass the 254 pixel boundary, the topdelta value becomes relative.