render the beta plasma ball sprites translucent
I know they were of course not translucent in the original beta
release, but we are using completely different plasma ball sprites
anyway - and they just look neat with some translucency added.

Alternative implementations would be to set this flag in
D_InitTables() instead, or to set it in a DEHACKED patch in
autoload/doom-all. What do you think?