add support for PC speaker sounds
This is an amalgamation of Chocolate Doom's pcsound/pcsound_sdl.c and
src/i_pcsound.c files.

This is an initial naive but working implementation. Instead of mixing
the sounds directly into the SDL_Mixer stream, the entire lumps get
synthetized as regular wave files with the correct target format,

Strictly speaking, PC speaker sounds do not support stereo separation,
pitch shifting and distance attenuation - and they don't adhere to the
sound channels concept. Question is if ths should be adapted or not?
It will pretty sure sound worse but more faithful. Not sure if it's
worth the effort.

Also, not sure if this should be kept as an easter egg or if it
deserves a promiment switch in the menu?