more robust fallback logic for music modules
The MIDI player fallback logic was to initialize `module 0 device 0`.
However, it is possible that this fallback also fails.

A scenario where this happens is when running on Linux with FluidSynth
but no soundfonts. In this case, FluidSynth will fail to initialize, the
fallback will try to use FluidSynth again, which will fail again, and
since the return code of `module->I_InitMusic` is not checked, the module
will stay uninitialized and will cause a crash a bit later.

To fix this, implement a more robust fallback logic which tries to
initialize each music module. This logic will always succeed since in
the worst case it will fallback to OPL, which currently never fails.

Fixes: 2f27fa739a1a ("fall back to `module 0 device 0` if the music module failed to initialize")