fix initialization with invalid video_display setting
If the `video_display` setting is invalid (for example, due to a
disconnected monitor), it was reset on `I_GetWindowPosition`, which
is called during video initialization.

However, commit dc4182d6d81f ("add exclusive fullscreen to the menu")
reordered the initialization logic. As a result, an invalid
`video_display` can be passed to `SDL_GetCurrentDisplayMode`, causing
Woof to exit before the `video_display` reset logic is run.

Commit cab179594848 ("i_video.c refactor, resize upscaled texture on
window resize") somewhat mitigates the problem, but an exit still
happens when `fullscreen = 1` and `exclusive_fullscreen = 1`.

Fix this by moving the `video_display` reset logic to a separate
function which is called at the very start of video initialization.