switch order of shaded and gold translation für submenu bar (#1420)
My sentiment is: If `v_lightest_color` is plain white, then this color
translated to gold may result in a very desaturated color, which then
gets shaded. I'd prefer if we translate a light shade of gray to gold
instead. Also, the colors in the HUD font are in general not full
bright, so translating from a moderate gray to gold may fit better.