DrawPatch refactoring (#2532)
* reorganized DrawPatch functions

moved around a lot of drawpatch function definitions to simplify the use
of DrawPatchInternal, which has a lot of potential inputs to it, placing
one specific function for each indiviual use of the on-screen patch
drawing behavior. horrid git diff aside, this also implements
translucent and translated drawing for custom end finale cast call
sequences.

* fix const-ness

* simplify deltaw calculation

* fix non-linux build issue

* general tweaks